Seguridad v4

+++ src/views/ABMUsuariosView.vue
@@ -11,7 +11,7 @@
             <th>Apellido</th>
             <th>Segundo Apellido</th>
             <th>Correo</th>
-            <th>Telefono</th>
+            <th>Teléfono</th>
             <th>Roles</th>
             <th>Estado</th>
             <th>Acciones</th>
@@ -24,11 +24,11 @@
             <td>{{ usuario.nombre }}</td>
             <td>{{ usuario.apellido }}</td>
             <td>{{ usuario.segundoApellido }}</td>
-            <td>{{ usuario.telefono }}</td>
             <td>{{ usuario.correo }}</td>
+            <td>{{ usuario.telefono }}</td>
             <td>
               <div v-for="rol in usuario.rol" :key="rol">
-                {{ rol }}
+                <div class="rol">{{ mapRol(rol) }}</div>
               </div>
             </td>
             <td>{{ usuario.estado }}</td>
@@ -53,7 +53,7 @@
               <th>Apellido</th>
               <th>Segundo Apellido</th>
               <th>Correo</th>
-              <th>Telefono</th>
+              <th>Teléfono</th>
               <th>Roles</th>
               <th>Estado</th>
             </tr>
@@ -61,7 +61,7 @@
           <tbody>
             <tr>
               <td>{{ usuarioSeleccionado.id }}</td>
-              <td>{{ usuarioSeleccionado.username }}</td>
+              <td><input v-model="usuarioSeleccionado.username" /></td>
               <td><input v-model="usuarioSeleccionado.nombre" /></td>
               <td><input v-model="usuarioSeleccionado.apellido" /></td>
               <td><input v-model="usuarioSeleccionado.segundoApellido" /></td>
@@ -71,7 +71,7 @@
                 <fieldset>
                   <label v-for="rol in roles" :key="rol">
                     <input type="checkbox" :value="rol" v-model="usuarioSeleccionado.rol" />
-                    {{ rol }}
+                    {{ mapRol(rol) }}
                   </label>
                 </fieldset>
               </td>
@@ -99,14 +99,45 @@
 <script>
 import { ref } from 'vue';
 import { getFirestore, collection, getDocs, doc, updateDoc, deleteDoc } from 'firebase/firestore';
-
+import { signOut } from "firebase/auth";
+import { auth } from "@/firebase"; // Configuración de Firebase
+import { mapState, mapActions } from 'vuex';
 export default {
   name: 'ABMUsuariosView',
+  computed: {
+    ...mapState(['isAuthenticated', 'user']),
+    username() {
+      return this.user?.username || '';
+    },
+    userInitials() {
+      if (!this.username) return '';
+      const names = this.username.split(' ');
+      const initials = names[0].charAt(0) + (names[1] ? names[1].charAt(0) : '');
+      return initials.toUpperCase();
+    }
+  },
   setup() {
     const usuarios = ref([]);
     const usuarioSeleccionado = ref(null);
-    const roles = ['usuario', 'Admin_ABM_Usuarios', 'Admin_ABM_Inventario', 'Admin_Ventas', 'Admin_Reportes'];
-    
+
+    // Mapeo de roles internos a nombres visuales
+    const rolesMap = {
+      'Usuario': 'Usuario',
+      'Admin_ABM_Usuarios': 'Administrador de Usuarios',
+      'Admin_ABM_Inventario': 'Administrador de Inventario',
+      'Admin_Ventas': 'Jefe de Ventas',
+      'Contador': 'Contador',
+      'Auditor': 'Auditor'
+     
+    };
+
+    const roles = ['Usuario','Admin_ABM_Usuarios', 'Admin_ABM_Inventario','Admin_Ventas', 'Contador', 'Auditor'];
+
+    // Función para mapear el rol a un valor visual
+    const mapRol = (rol) => {
+      return rolesMap[rol] || rol; // Retorna el nombre visual, o el rol original si no está en el mapa
+    };
+
     const fetchUsuarios = async () => {
       const db = getFirestore();
       const usuariosCollection = collection(db, 'users');
@@ -137,13 +168,21 @@ export default {
       await updateDoc(usuarioDoc, usuarioSeleccionado.value);
       usuarioSeleccionado.value = null; // Limpiar la selección
       fetchUsuarios(); // Volver a cargar la lista
+
+      // Refresca la página
+      window.location.reload();
     };
 
     const eliminarUsuario = async (id) => {
-      const db = getFirestore();
-      const usuarioDoc = doc(db, 'users', id);
-      await deleteDoc(usuarioDoc);
-      fetchUsuarios(); // Volver a cargar la lista
+      const confirmacion = confirm('¿Estás seguro de que deseas eliminar este usuario?');
+      if (confirmacion) {
+        const db = getFirestore();
+        const usuarioDoc = doc(db, 'users', id);
+        await deleteDoc(usuarioDoc);
+        fetchUsuarios(); // Volver a cargar la lista
+      } else {
+        alert('Eliminación cancelada.');
+      }
     };
 
     // Cargar usuarios al inicio
@@ -153,6 +192,7 @@ export default {
       usuarios,
       usuarioSeleccionado,
       roles,
+      mapRol,
       fetchUsuarios,
       editarUsuario,
       guardarCambios,
@@ -164,6 +204,7 @@ export default {
 
 <style scoped>
 /* Estilos de la tabla y formulario */
+
 .tabla-contenedor {
   overflow-x: auto;
 }
@@ -171,6 +212,7 @@ export default {
 table {
   width: 100%;
   border-collapse: collapse;
+  background-color:antiquewhite;
 }
 
 th, td {
@@ -184,7 +226,15 @@ td {
 th {
   background-color: #FF9800;
 }
-
+.rol{
+  text-align: center;
+  margin: 5px;
+  color: aliceblue;
+  padding: 3px;
+  background: #1eb039;
+  border: 2px solid #858585;
+  border-radius: 10px;
+}
 button {
   display: inline-block; /* Hace que los botones se muestren en la misma línea */
   padding: 8px;
@@ -222,7 +272,15 @@ form {
   margin-top: 20px;
   border: 1.5px solid #ccc;
 }
+form button{
+  background-color: blue;
+  color: white;
+}
 
+form button:hover{
+  background-color: rgb(77, 77, 247);
+  color: white;
+}
 fieldset {
   border: none;
   margin: 10px 0;
